draw normal edit

+++ src/viewer.hpp
@@ -12,8 +12,7 @@ private:
 	const bool showNormals;
 
 	const rl::Vector3 x_axis{ 1,0,0 };
-	const float axis_rotation{ - nc::constants::pi / 2 };
-	//const float axis_rotation{ 0 };
+	const float axis_rotation{ (double)(- nc::constants::pi / 2) };
 
 	MultiMesh* currentMesh;
 	const std::vector<std::array<rl::Vector3, 2>> currentMeshLines;
